WIP rework meal plans

+++ src/FoodStuffs.Model/Events/MealPlans/SaveMealPlanRequestPantryIngredient.cs
@@ -2,4 +2,4 @@
 
 public record SaveMealPlanRequestPantryIngredient(
     string Name,
-    decimal Quantity);
+    int Quantity);
